[chore][pkg/stanza] Cleanup uri parser operator files
