+++ pkg/stanza/operator/parser/uri/parser_test.go
